Fixed TypeError in QtScheduler
Fixes #687.

+++ apscheduler/schedulers/qt.py
@@ -33,7 +33,7 @@ class QtScheduler(BaseScheduler):
     def _start_timer(self, wait_seconds):
         self._stop_timer()
         if wait_seconds is not None:
-            wait_time = min(wait_seconds * 1000, 2147483647)
+            wait_time = min(int(wait_seconds * 1000), 2147483647)
             self._timer = QTimer.singleShot(wait_time, self._process_jobs)
 
     def _stop_timer(self):
